feat: add core library and update references
