+++ Windirstat.Core/S3Scanner.cs
@@ -1,25 +1,14 @@
 using Amazon.S3;
 using Amazon.S3.Model;
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Windirstat.Core.Models;
 
-namespace windirstat_s3.Services;
-
-public class FolderNode
-{
-    public string Name { get; }
-    public long Size { get; set; }
-    public long FileCount { get; set; }
-    public long OwnSize { get; set; }
-    public DateTime LastModified { get; set; }
-    public Dictionary<string, FolderNode> Children { get; } = new();
-    public Dictionary<string, ExtensionInfo> Extensions { get; } = new();
-
-    public FolderNode(string name)
-    {
-        Name = name;
-    }
-}
+namespace Windirstat.Core;
 
 public class S3Scanner
 {
