+++ s3_config_cli/s3_config_cli.csproj
@@ -5,4 +5,8 @@
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\\Windirstat.Core\\Windirstat.Core.csproj" />
+  </ItemGroup>
 </Project>
