+++ windirstat_s3/windirstat_s3.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="AWSSDK.S3" Version="4.0.6.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\\Windirstat.Core\\Windirstat.Core.csproj" />
+  </ItemGroup>
+
 </Project>
